+++ test/unit/expirationd_stats_test.lua
@@ -11,8 +11,9 @@ local g = t.group('expirationd_stats', {
 })
 
 g.before_each({index_type = 'TREE'}, function(cg)
-    t.skip_if(cg.params.engine == 'vinyl' and helpers.vinyl_is_broken(),
-        'Blocked by https://github.com/tarantool/tarantool/issues/6448')
+    t.skip_if(cg.params.engine == 'vinyl' and not helpers.vinyl_is_supported(),
+        'Blocked by https://github.com/tarantool/tarantool/issues/6448 on ' ..
+		'this Tarantool version')
     g.space = helpers.create_space_with_tree_index(cg.params.engine)
 end)